refactor(grouping package): grouping package 에 OrderTableGrouped/UngroupedEventHandler 이동하여 의존성 분리
grouping package 에 OrderTableGrouped/UngroupedEventHandler 이동하여 의존성을 분리한다.
- OrderTableGroupedEventHandler, OrderTableUngroupedEventHandler : grouping package 로 이동
- TableGroup, TableGroupGroupedEvent, TableGroupUngroupedEvent, TableGroupServiceTest : Event 파라미터를 TableGroup -> tableGroupId 로 변경
- after.png : After 이미지 수정
