+++ src/main/java/kitchenpos/application/grouping/OrderTableGroupedEventHandler.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.table;
+package kitchenpos.application.grouping;
 
 import java.util.List;
 
@@ -10,7 +10,6 @@ import org.springframework.transaction.annotation.Transactional;
 import kitchenpos.domain.table.OrderTable;
 import kitchenpos.domain.table.OrderTableRepository;
 import kitchenpos.domain.tablegroup.event.TableGroupGroupedEvent;
-import kitchenpos.domain.tablegroup.TableGroup;
 
 @Component
 public class OrderTableGroupedEventHandler {
@@ -25,9 +24,7 @@ public class OrderTableGroupedEventHandler {
     @EventListener
     @Transactional
     public void handle(TableGroupGroupedEvent event) {
-        TableGroup tableGroup = event.getTableGroup();
         List<OrderTable> orderTables = orderTableRepository.findAllByIdIn(event.getOrderTableIds());
-
-        orderTables.forEach(orderTable -> orderTable.alignTableGroup(tableGroup.getId()));
+        orderTables.forEach(orderTable -> orderTable.alignTableGroup(event.getTableGroupId()));
     }
 }
