+++ src/main/java/kitchenpos/application/grouping/OrderTableUngroupedEventHandler.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.table;
+package kitchenpos.application.grouping;
 
 import java.util.List;
 
@@ -9,7 +9,6 @@ import org.springframework.transaction.annotation.Transactional;
 import kitchenpos.domain.table.OrderTable;
 import kitchenpos.domain.table.OrderTableRepository;
 import kitchenpos.domain.tablegroup.event.TableGroupUngroupedEvent;
-import kitchenpos.domain.tablegroup.TableGroup;
 
 @Component
 public class OrderTableUngroupedEventHandler {
@@ -23,9 +22,7 @@ public class OrderTableUngroupedEventHandler {
     @EventListener
     @Transactional
     public void handle(TableGroupUngroupedEvent event) {
-        TableGroup tableGroup = event.getTableGroup();
-        List<OrderTable> orderTables = orderTableRepository.findAllByTableGroupId(tableGroup.getId());
-
+        List<OrderTable> orderTables = orderTableRepository.findAllByTableGroupId(event.getTableGroupId());
         orderTables.forEach(OrderTable::ungroup);
     }
 }
