+++ src/main/java/kitchenpos/domain/tablegroup/event/TableGroupGroupedEvent.java
@@ -2,20 +2,18 @@ package kitchenpos.domain.tablegroup.event;
 
 import java.util.List;
 
-import kitchenpos.domain.tablegroup.TableGroup;
-
 public class TableGroupGroupedEvent {
 
-    private TableGroup tableGroup;
+    private Long tableGroupId;
     private List<Long> orderTableIds;
 
-    public TableGroupGroupedEvent(TableGroup tableGroup, List<Long> orderTableIds) {
-        this.tableGroup = tableGroup;
+    public TableGroupGroupedEvent(Long tableGroupId, List<Long> orderTableIds) {
+        this.tableGroupId = tableGroupId;
         this.orderTableIds = orderTableIds;
     }
 
-    public TableGroup getTableGroup() {
-        return tableGroup;
+    public Long getTableGroupId() {
+        return tableGroupId;
     }
 
     public List<Long> getOrderTableIds() {
